Validation for extension to be .json
We have added a validation to check if the file's extension is .json before uploading it to the database, this was done in the frontend part to prevent any errors in the backend as well. Note that importing materialUi is needed for the way we handled this issue and using "npm install @material-ui/lab" to install it on the project folder directory is also needed.