Cleaned up Profile forms and password forms.
Implemented forms in a similar manner to the form for entering dataset information. Users can toggle between changing their profile settings and changing their password. Currently not connected to the backend. Additionally, does not match the mockups due to I not knowing of any other settings that need to be implemented.