Moved calculationh of graph extremities to GraphView
This portion of the code was moved because of a bug that occured when removing a graph that had more extreme points than a previous dataset. Before, the max/min would remain as the max/min of the previous dataset/ Now, as it is calculated whenever changes to the loaded dataset are done, this issue should no longer exist.